backend/requirements.txt
Django>=5.0,<6.0
gunicorn>=21.2,<23.0
django-cors-headers>=4.3,<5.0
cryptography==46.0.5
boto3==1.42.49
python-dotenv==1.2.1